+++ js/main.js
@@ -3,6 +3,7 @@ import { initEditorTab, navigateToLine } from "./editorTab.js";
 import { initViewerTab } from "./viewerTab.js";
 import { initPreviewTab } from "./previewTab.js";
 import { initSearch } from "./search.js";
+import { initProjectFile } from "./projectFile.js";
 
 window.navigateToLine = navigateToLine;
 
@@ -12,5 +13,6 @@ document.addEventListener("DOMContentLoaded", () => {
   initViewerTab();
   initPreviewTab();
   initSearch();
+  initProjectFile();
   runStartupFlow();
 });